Allow server-side email filtering
Emails can be filtered to restrict the accepted emails. The user
has to provide a function doing the filering, possibly in an
asynchronous manner. The function receives two callbacks, one to
call if the email is valid (nothing to provide) and one to call
if the email is invalid (with the rejection reason to provide as
argument).